backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Validation and serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0

# Push notifications (APNs)
aioapns==3.1

# HTTP client for external APIs
httpx==0.26.0

# Machine learning for probability model
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4

# Background tasks
apscheduler==3.10.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
aiosqlite==0.19.0
greenlet==3.0.3

# Utilities
python-dotenv==1.0.0
